requirements.txt
# Web Framework & API
fastapi==0.109.2
uvicorn[standard]==0.27.1
gunicorn==21.2.0
python-multipart==0.0.9
requests==2.31.0
httpx==0.27.0
aiofiles==23.2.1

# Database
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.1

# Configuration
pydantic==2.6.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.2

# Document Processing
PyPDF2==3.0.1
pdfminer.six==20231228
python-docx==1.1.0
python-pptx==0.6.23
Pillow==10.2.0
beautifulsoup4==4.12.3
lxml==5.1.0

# OCR & Transcription
pytesseract==0.3.10
openai-whisper==20231117
opencv-python-headless==4.9.0.80

# AI & Embeddings
anthropic==0.18.1
openai==1.12.0
sentence-transformers==2.3.1
langchain==0.1.5
langchain-community==0.0.17
langchain-openai==0.0.5
langchain-anthropic==0.1.1
numpy==1.26.4
transformers==4.37.2

# Vector Databases
faiss-cpu==1.8.0
pinecone-client==3.1.0
weaviate-client==3.26.0

# Integrations
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
notion-client==2.2.1

# Utilities
tqdm==4.66.2
schedule==1.2.1
nltk==3.8.1
spacy==3.7.4

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0supabase==2.3.0
pgvector==0.2.5
